+++ app/assets/stylesheets/application.css.scss
@@ -229,6 +229,12 @@ td {
 	    color: white;
 	    background: black;
 	  }
+	  .no-stats{
+	    background: black;
+	    color: white;
+	    border: 2px solid black;
+	    border-radius: 5px;
+	  }
 
 	}
 }
