+++ app/controllers/profiles_controller.rb
@@ -1,6 +1,5 @@
 require 'lol'
 require 'certified'
-require 'ostruct'
 class ProfilesController < ApplicationController
   attr_accessor :summonerObj, :statsObj, :champsReq, :summonerReq, :statsReq
   before_action :verify_summoner_name_and_stats, only: [:create]
@@ -31,6 +30,23 @@ class ProfilesController < ApplicationController
     end
   end
 
+  def search
+    @name = params[:friend]
+    if params[:friend]
+      create_search_connection
+      begin
+        @friend = @summonerReq.by_name(params[:friend].downcase).first
+        get_summoner_ranked_stats(@friend.id)
+        
+      rescue Lol::NotFound => e
+        if e.message == '404 Not Found'
+          return
+        end
+      end
+      get_search_stats
+    end
+  end
+
   private
 
   def user_params
@@ -47,7 +63,15 @@ class ProfilesController < ApplicationController
   def create_champ_connection
     @champsReq = Lol::StaticRequest.new Rails.application.secrets.sulai_api_key, "na"
   end
-
+  def get_search_stats
+    get_most_played_champ
+    @mostPlayedChampObj = @mostPlayedChampObj[0]
+    @mpChamp = @champsReq.get('champion', @mostPlayedChampObj.id)
+    get_highest_KDR_champ
+    @hkdrChamp = @champsReq.get('champion', @highestKillDeathRatioChampObj[0].id)
+    get_highest_WLR_champ
+    @hwlrChamp = @champsReq.get('champion', @highestWinLossRatioChampObj[0].id)
+  end
 
   def verify_summoner_name_and_stats
 #So far, only three requests are being used. Stats Request, Static Request and Summoner Request.
@@ -79,7 +103,7 @@ class ProfilesController < ApplicationController
   def get_summoner
     create_summoner_connection
     @summonerObj = @summonerReq.by_name(User.find(params[:id]).name).first
-    get_summoner_ranked_stats
+    get_summoner_ranked_stats(@summonerObj.id)
     create_champ_connection
     get_most_played_champ
     get_overall_WLR
@@ -88,28 +112,21 @@ class ProfilesController < ApplicationController
     get_highest_WLR_champ
   end
   
-  #Static Request is used to retrieve info on any of champion, item, mastery, rune summoner_spell
-  
-
-  def get_summoner_ranked_stats
+  def get_summoner_ranked_stats(search_id)
     create_stats_connection
-    @statsObj = statsReq.ranked(@summonerObj.id, extra = {})
-    #Since the champion with id = 0 always returns 404 error and is used for error checking, I am deleting it from the array.
-    @statsObj.champions.delete_if{ |h| h.id == 0 }
+    @statsObj = statsReq.ranked(search_id, extra = {})
+    #Since the champion with id = 0 always returns 404 error and is used for error checking, it is being deleted from array.
+    if @statsObj
+      @statsObj.champions.delete_if{ |h| h.id == 0 }
+    end
   end
 
   def get_most_played_champ
-    #Most played champion is returned by sorting by total sessions played descending and returning second champion.
-    #First champion does not exist - refer below
-    #Most summoners I have tried have ID = 0 for champions with the highest total session played.
-    #However, I had trouble retrieving this champion and after researching for a while, I found that champion with ID = 0 does not exist and
-    #is only used to deal with exceptions. Linked here - https://market.mashape.com/community/elophant-league-of-legends
     @mostPlayedChampObj = @statsObj.champions.sort_by{|x| -x.stats.total_sessions_played}
     @mostPlayedChampObj = @mostPlayedChampObj.paginate(:page => params[:page], :per_page => 5)
   end
 
   def get_overall_WLR
-    #OverallWLR
     overallLosses = 0
     overallWins = 0                                              
     @statsObj.champions.each do |x|
@@ -124,7 +141,6 @@ class ProfilesController < ApplicationController
   end
 
   def get_overall_KDR
-    #OverallKDR
     overallKills = 0
     overallDeaths = 0                                              
     @statsObj.champions.each do |x|
@@ -140,12 +156,6 @@ class ProfilesController < ApplicationController
   end
   
   def get_highest_KDR_champ
-    #Champ with highest KDR (KDR here means Champions Killed/Champion's Death Ratio
-    #There are other Kill stats that I didn't consider valuable enough to include
-    #After doing some research, it also appears that KDR is subjective: http://leagueoflegends.wikia.com/wiki/Kill_to_Death_Ratio
-    #. This is tricky, as some Champs might have zero deaths. In the case of zero death, division is
-    #made by zero to avoid divide by zero error. Return first element of descending array.
-    #KDR for a champion is calculated per session instead of using kill values that might not be reflective of per session values.
     @highestKillDeathRatioChampObj = @statsObj.champions.sort_by{|x|  if x.stats.total_deaths_per_session >  0 then 
                                                                         -((x.stats.most_champion_kills_per_session.to_f)/(x.stats.total_deaths_per_session))
                                                                       else 
@@ -166,5 +176,8 @@ class ProfilesController < ApplicationController
     }
     @highestWinLossRatioChampObj = @highestWinLossRatioChampObj.paginate(:page => params[:page], :per_page => 5)
   end
-
+  def create_search_connection
+    create_summoner_connection
+    create_champ_connection
+  end
 end
