--- app/views/common/search.html.erb
@@ -1,75 +1,0 @@
-<div class = "search-page" id="friend-lookup">
-  
-  <div class = "col-md-4 col-md-offset-4">
-  	<div class="table-responsive">
-		  <table class="table search-table" >
-		    <thead>
-		    	<tr>
-		    		<th class="header search-header">Summoner Result</th>
-		    	</tr>
-		    </thead>
-		    <tbody>
-		      <tr>
-		    			<td>
-		    			  <% if @friend %>
-			    		    <div id="friend-lookup-results" class="row well results-block">
-			    		      <div class="col-sm-1">
-			    		        <% if @friendStatsObj %>
-			    		      	  <%= link_to image_tag("icons/#{@thisChampObj[:name].delete(' ')}" + "Square.png"), champion_path(@thisChampObj.id)  %>
-			    		        <% end %>
-			    		      </div>
-			    		      <div class="col-sm-9 col-sm-offset-2">
-			    		      	<strong>Summoner Name: <%= @name.capitalize %></strong><br>
-			    		        <strong>Summoner Level: <%= @friend.summoner_level %></strong><br>
-	                    <strong>Last Active: <%= @friend.revision_date.to_s.slice(0,10) %></strong><br>
-			    		      </div>
-			    		    </div>
-			    		  <% else %>
-			    		    <div id="friend-lookup-errors" class="well results-block">
-			    		      <strong><%= @name.capitalize %> is not recognized as a League of Legends player.</strong><br>
-			    		      <strong> Verify that summoner is active and try again.</strong><br>
-			    		    </div>
-			    		  <% end %>
-		    		  </td>
-		      </tr>
-		    </tbody>
-		    <tfoot>
-		    	<tr>
-	    			<td>
-		    	  	  <div class="col-sm-12 btn-group btn-group-justified" role="group" aria-label="...">
-								  <% if @friend %>
-									  <% if logged_in? %>
-				    			    <% @user = User.find_summoner_by_name(@name) %>
-				    			    <% if @user %>
-				    			      <div class="btn-group" role="group">
-											    <button type="button" class="btn btn-success">Add Friend</button>
-											  </div>
-											<% else %>
-	                      <div class="btn-group" role="group">
-											    <button type="button" class="btn btn-success">Invite Friend</button>
-											  </div>
-											<% end %>
-											<div class="btn-group" role="group">
-										    <button type="button" class="btn btn-default">View Profile</button>
-										  </div>
-	                  <% else %>
-	                    <div class="btn-group" role="group">
-											  <button type="button" class="btn btn-success">Login</button>
-											</div>
-											<div class="btn-group" role="group">
-											  <button type="button" class="btn btn-primary">Sign Up</button>
-											</div>
-				    			  <% end %>
-				    			<% end %>
-		              <div class="btn-group" role="group">
-								    <button type="button" class="btn btn-dismiss">Dismiss</button>
-								  </div>
-								</div>
-	    	  	</td>
-		    	</tr>
-		    </tfoot>
-		  </table>
-    </div>
-  </div>
-</div>
-
