+++ app/views/profiles/search.html.erb
@@ -1,0 +1,98 @@
+<div class = "search-page" id="friend-lookup">
+  
+  <div class = "col-md-6 col-md-offset-3">
+  	<div class="table-responsive">
+		  <table class="table search-table" >
+		    <thead>
+		    	<tr>
+		    		<th class="header search-header">Summoner Search Result</th>
+		    	</tr>
+		    </thead>
+		    <tbody>
+		      <tr>
+		    			<td>
+		    			  <% if @friend %>
+			    		    <div id="friend-lookup-results" class="row well results-block">
+			    		      <div class="col-sm-5">
+			    		      	<strong>Summoner Name: <%= @name.capitalize %></strong><br>
+			    		        <strong>Summoner Level: <%= @friend.summoner_level %></strong><br>
+	                    <strong>Last Active: <%= @friend.revision_date.to_s.slice(0,10) %></strong><br>
+			    		      </div>
+			    		      <% if @statsObj %>
+			    		      	<div class="col-sm-5">
+			    		          <strong>Most Played Champ: <%= "#{@mpChamp[:name]}"%> </strong><%= " - (#{@mostPlayedChampObj.stats.total_sessions_played}x)"%> <br>
+			    		          <strong>Highest KDR Champ: <%= "#{@hkdrChamp[:name]}"%> </strong> 
+			    		                                    <% if @highestKillDeathRatioChampObj[0].stats.total_deaths_per_session >  0 %>
+	                                                  <%= " - (#{((@highestKillDeathRatioChampObj[0].stats.most_champion_kills_per_session.to_f)/(@highestKillDeathRatioChampObj[0].stats.total_deaths_per_session)).round(2)})"%>
+	                                                <%else%>
+	                                                  <%= " - (#{(@highestKillDeathRatioChampObj[0].stats.most_champion_kills_per_session_f.round(2))})"%>
+	                                                <%end%><br>
+	                      <strong>Highest WLR Champ: <%= "#{@hwlrChamp[:name]}"%> </strong> 
+			    		                                    <% if @highestWinLossRatioChampObj[0].stats.total_deaths_per_session >  0 %>
+	                                                  <%= " - (#{((@highestWinLossRatioChampObj[0].stats.most_champion_kills_per_session.to_f)/(@highestWinLossRatioChampObj[0].stats.total_deaths_per_session)).round(2)})"%>
+	                                                <%else%>
+	                                                  <%= " - (#{(@highestWinLossRatioChampObj[0].stats.most_champion_kills_per_session_f.round(2))})"%>
+	                                                <%end%><br>
+			    		        </div>
+			    		        <div class="col-sm-1">
+			    		        <!--While this displays a dynamic picture(most played champ), we might need to have users choose their preferred display pictures
+			    		          when searched for or on their profile. This could be done when editing profiles.-->
+				    		        <%= link_to image_tag("icons/#{@mpChamp[:name].delete(' ')}Square.png"), champion_path(@mpChamp.id)  %>
+				    		      </div>
+			    		      <% else %>
+			    		        <div class="col-sm-6 no-stats">
+			    		          <strong>No stats found for summoner.</strong><br>
+				    		        <strong>Either summoner is inactive or</strong><br>
+		                    <strong>there are no stats to display.</strong><br>
+			    		        </div>
+			    		      <% end %>
+			    		    </div>
+			    		  <% else %>
+			    		    <div id="friend-lookup-errors" class="well results-block">
+			    		      <strong><%= @name.capitalize %> is not a recognized League of Legends player.</strong><br>
+			    		      <strong> Verify that <%= @name.capitalize %> is a valid summoner name and try again.</strong><br>
+			    		    </div>
+			    		  <% end %>
+		    		  </td>
+		      </tr>
+		    </tbody>
+		    <tfoot>
+		    	<tr>
+	    			<td>
+		    	  	  <div class="col-sm-12 btn-group btn-group-justified" role="group" aria-label="...">
+								  <% if @statsObj %>
+									  <% if logged_in? %>
+				    			    <% @user = User.find_summoner_by_name(@name) %>
+				    			    <% if @user %>
+				    			      <div class="btn-group" role="group">
+				    			        <%= link_to "Add Friend".html_safe, profile_path(@current_user.id), :class =>"btn btn-success" %>
+											  </div>
+											  <div class="btn-group" role="group">
+											    <%= link_to "View Profile".html_safe, profile_path(@user.id), :class =>"btn btn-default" %>
+											  </div>
+											<% else %>
+	                      <div class="btn-group" role="group">
+											    <%= link_to "Invite Friend".html_safe, root_path, :class =>"btn btn-success" %>
+											  </div>
+											<% end %>
+										<% else %>
+	                    <div class="btn-group" role="group">
+											  <%= link_to "Login".html_safe, login_path, :class =>"btn btn-success" %>
+											</div>
+											<div class="btn-group" role="group">
+											  <%= link_to "Sign Up".html_safe, new_profile_path, :class =>"btn btn-primary" %>
+											</div>
+				    			  <% end %>
+				    			<% end %>
+		              <div class="btn-group" role="group">
+								    <%= link_to "Dismiss".html_safe, root_path, :class =>"btn btn-danger" %>
+								  </div>
+								</div>
+	    	  	</td>
+		    	</tr>
+		    </tfoot>
+		  </table>
+    </div>
+  </div>
+</div>
+
